Preliminary support padding, log on panic.
This adds preliminary support for data padding by adding another layer
of encapsulation inside each AEAD frame containing a type and length.
For now, data is still sent unpadded, but the infrastructure for
supporting it is mostly there.

Additionally, use log.Panic[f]() instead of panic through out the code
so that some panics are logged.