Change the length field obfscation.
Instead of including the previous secretbox in the input when
calculating the  SipHash-2-4 digest used to generate the obfuscation
mask, use only the nonce.  This is significantly faster, and if someone
breaks obfs4 by exploiting the low amount of input entropy between each
invocation (a counter incrementing by 1), I hope they publish the
attack on the PRF.

This breaks wire protocol compatibility.