Change the import path for go.net.
The Go developers decided to move the go.net repository to
golang.org/x/net, and also to transition from hg to git.  This wasn't
changed when the go.crypto imports were since the 'proxy' component
doesn't have imports that break, so the old code still works.

While the change here is simple (just update the import location), this
affects packagers as it now expects the updated package.  Sorry for the
inconveneince, I blame the Go people, and myself for not just doing
this along with the go.crypto changes.